Start refactoring communications system
