+++ app/proguard-gradle.txt
@@ -52,7 +52,7 @@
 }
 
 # This class is parsed/marshalled by Gson
--keep class com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticDevice { *; }
+-keep class com.paulmandal.atak.forwarder.comm.meshtastic.MeshtasticDevice { *; }
 
 # These classes use Parcelable
 -keep class com.geeksville.mesh.** { *; }
