+++ app/src/main/java/com/paulmandal/atak/forwarder/comm/meshtastic/MeshtasticDevice.java
@@ -1,4 +1,4 @@
-package com.paulmandal.atak.forwarder.comm.commhardware.meshtastic;
+package com.paulmandal.atak.forwarder.comm.meshtastic;
 
 import androidx.annotation.NonNull;
