+++ app/src/main/java/com/paulmandal/atak/forwarder/comm/refactor/MeshtasticDeviceConfigurer.java
@@ -1,4 +1,4 @@
-package com.paulmandal.atak.forwarder.comm.commhardware.meshtastic;
+package com.paulmandal.atak.forwarder.comm.refactor;
 
 import android.os.RemoteException;
 import android.util.Log;
