+++ app/src/main/java/com/paulmandal/atak/forwarder/comm/refactor/MeshtasticDeviceSwitcher.java
@@ -1,4 +1,4 @@
-package com.paulmandal.atak.forwarder.comm.commhardware.meshtastic;
+package com.paulmandal.atak.forwarder.comm.refactor;
 
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
@@ -12,6 +12,7 @@ import android.util.Log;
 
 import com.geeksville.mesh.IMeshService;
 import com.paulmandal.atak.forwarder.Config;
+import com.paulmandal.atak.forwarder.comm.meshtastic.MeshtasticDevice;
 
 import java.util.ArrayList;
 import java.util.HashMap;
