+++ app/src/main/java/com/paulmandal/atak/forwarder/factories/InternalComponentsFactory.java
@@ -1,0 +1,7 @@
+package com.paulmandal.atak.forwarder.factories;
+
+public class InternalComponentsFactory {
+    public void createInternalComponents() {
+
+    }
+}
