+++ app/src/main/java/com/paulmandal/atak/forwarder/plugin/ui/ForwarderMarkerIconWidget.java
@@ -15,7 +15,7 @@ import com.atakmap.android.widgets.RootLayoutWidget;
 import com.atakmap.coremap.maps.assets.Icon;
 import com.paulmandal.atak.forwarder.R;
 import com.paulmandal.atak.forwarder.comm.commhardware.CommHardware;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticCommHardware;
+import com.paulmandal.atak.forwarder.comm.refactor.MeshtasticCommHardware;
 import com.paulmandal.atak.forwarder.plugin.Destroyable;
 
 import java.util.List;
