+++ app/src/main/java/com/paulmandal/atak/forwarder/plugin/ui/settings/DevicesList.java
@@ -8,7 +8,7 @@ import android.hardware.usb.UsbDevice;
 import android.hardware.usb.UsbManager;
 
 import com.paulmandal.atak.forwarder.Config;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticDevice;
+import com.paulmandal.atak.forwarder.comm.meshtastic.MeshtasticDevice;
 
 import java.util.ArrayList;
 import java.util.Collection;
