+++ app/src/main/java/com/paulmandal/atak/forwarder/plugin/ui/settings/MainButtons.java
@@ -5,7 +5,7 @@ import android.preference.Preference;
 import com.atakmap.android.gui.PanListPreference;
 import com.google.gson.Gson;
 import com.paulmandal.atak.forwarder.Config;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticDevice;
+import com.paulmandal.atak.forwarder.comm.meshtastic.MeshtasticDevice;
 
 import java.util.List;
