+++ app/src/main/java/com/paulmandal/atak/forwarder/plugin/ui/settings/TrackerButtons.java
@@ -23,9 +23,9 @@ import com.atakmap.android.gui.PluginSpinner;
 import com.geeksville.mesh.MeshProtos;
 import com.google.gson.Gson;
 import com.paulmandal.atak.forwarder.R;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticCommHardware;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticDeviceSwitcher;
-import com.paulmandal.atak.forwarder.comm.commhardware.meshtastic.MeshtasticDevice;
+import com.paulmandal.atak.forwarder.comm.refactor.MeshtasticCommHardware;
+import com.paulmandal.atak.forwarder.comm.refactor.MeshtasticDeviceSwitcher;
+import com.paulmandal.atak.forwarder.comm.meshtastic.MeshtasticDevice;
 import com.paulmandal.atak.forwarder.tracker.MeshtasticTrackerConfigurator;
 import com.paulmandal.atak.forwarder.preferences.PreferencesDefaults;
 import com.paulmandal.atak.forwarder.preferences.PreferencesKeys;
